+++ node/bo/assets/styles/framework/_popup.scss
@@ -2,70 +2,41 @@
 
 .v-overlay-container {
   .v-overlay__scrim {
-    opacity: 0.34;
+    opacity: 0.42;
   }
   .v-dialog {
     .v-overlay__content {
       min-width: 300px;
+      max-width: 1600px;
       .v-card,
       .v-sheet {
-        border-radius: 8px;
+        border-radius: 13px;
       }
 
       .v-card {
+        max-height: 700px;
         .v-card-title {
           position: relative;
-          font-family: 'Pretendard';
-          font-weight: 700;
-          font-size: pxToRem(18);
-          color: #000;
-          padding: 16px 20px;
+          font-family: 'Pretendard, sans-serif';
+          font-weight: 600;
+          font-size: pxToRem(20);
+          color: black;
+          padding: 30px 30px 20px 30px;
           line-height: 24px;
 
           .closeBtn {
             position: absolute;
-            right: 18px;
-            top: 18px;
-            width: 18px;
+            top: 30px;
+            right: 30px;
+            width: 20px;
             height: 20px;
             cursor: pointer;
-            transform: rotateZ(0);
-            transform-origin: center center;
-            transition: all 0.4s ease-out;
-            &:hover {
-              transform: rotateZ(360deg);
-            }
-            &:before {
-              content: '';
-              display: block;
-              position: absolute;
-              top: 0px;
-              left: 8px;
-              width: 1px;
-              height: 20px;
-              background-color: #444;
-              -webkit-transform: rotate(45deg);
-              transform: rotate(45deg);
-            }
-            &:after {
-              content: '';
-              display: block;
-              position: absolute;
-              top: 0px;
-              left: 8px;
-              width: 1px;
-              height: 20px;
-              background-color: #444;
-              -webkit-transform: rotate(135deg);
-              transform: rotate(135deg);
-            }
+            background: url(/assets/images/svg/ico_close.svg) no-repeat center center;
           }
         }
         .v-card-text {
-          padding: 20px;
-          border-top: 1px solid #e0e0e0;
-          border-bottom: 1px solid #e0e0e0;
-          min-height: 80px;
+          min-height: 100px;
+          padding: 0px 30px 25px 30px;
         }
         .v-card-actions {
           padding: 20px;
